Refactor: Separating code, fixing stylizations, adding format functions, removing comments
