+++ src/app/(dashboard)/dashboard/(home)/content.tsx
@@ -1,13 +1,11 @@
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card'
+import { formatBRL } from '@/lib/format'
+
 import { Separator } from '@/components/ui/separator'
+import { Card, CardHeader } from '@/components/ui/card'
 import { Content, ContentHeader } from '@/components/content'
+
+import { Large, Small } from '@/components/typography/texts'
+import { Heading } from '@/pegasus/heading'
 
 export function DashboardContent() {
   return (
@@ -19,34 +17,25 @@ export function DashboardContent() {
 
       <Separator className='mb-4' />
 
-      <div className='flex gap-4'>
-        <Card>
-          <CardHeader>
-            <CardTitle>Card Title</CardTitle>
-            <CardDescription>Card Description</CardDescription>
-          </CardHeader>
-          <CardContent>Card Content</CardContent>
-          <CardFooter>Card Footer</CardFooter>
-        </Card>
+      <section className='space-y-2.5'>
+        <Heading variant='h4'>Últimos 30 dias</Heading>
 
-        <Card>
-          <CardHeader>
-            <CardTitle>Card Title</CardTitle>
-            <CardDescription>Card Description</CardDescription>
-          </CardHeader>
-          <CardContent>Card Content</CardContent>
-          <CardFooter>Card Footer</CardFooter>
-        </Card>
+        <div className='flex gap-2.5'>
+          <Card className='flex flex-1 items-center justify-center shadow-md'>
+            <CardHeader className='flex flex-col items-center'>
+              <Large>Receita esperada</Large>
+              <Small>{formatBRL(1200)} em 25 pedidos pendentes</Small>
+            </CardHeader>
+          </Card>
 
-        <Card>
-          <CardHeader>
-            <CardTitle>Card Title</CardTitle>
-            <CardDescription>Card Description</CardDescription>
-          </CardHeader>
-          <CardContent>Card Content</CardContent>
-          <CardFooter>Card Footer</CardFooter>
-        </Card>
-      </div>
+          <Card className='flex flex-1 items-center justify-center shadow-md'>
+            <CardHeader className='flex flex-col items-center'>
+              <Large>Receita confirmada</Large>
+              <Small>{formatBRL(1200)} em 12 pedidos completos</Small>
+            </CardHeader>
+          </Card>
+        </div>
+      </section>
     </Content>
   )
 }
