+++ src/app/(dashboard)/dashboard/catalog/produtos/(table)/products-table.tsx
@@ -4,13 +4,13 @@ import * as React from 'react'
 
 import { Product } from '@/payload/payload-types'
 
+import { getColumns } from './columns'
 import { type ColumnDef } from '@tanstack/react-table'
 
 import { DataTable } from '@/components/table/data-table'
 import { useDataTable } from '@/components/table/hooks/use-data-table'
 import { NewProductDialog } from '../_components/new-product-dialog'
 
-import { getColumns } from './columns'
 import { getProducts } from '../_logic/queries'
 
 interface ProductsTableProps {
