+++ src/app/(dashboard)/dashboard/catalog/produtos/_components/new-product-form.tsx
@@ -3,20 +3,15 @@
 import { useEffect, useState } from 'react'
 
 import { z } from 'zod'
-import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
+import { useForm, useFormState } from 'react-hook-form'
 
 import { Attribute, Category } from '@/payload/payload-types'
 
 import { toast } from 'sonner'
 import { AnimatePresence, motion } from 'framer-motion'
 
-import {
-  filterAttributesByName,
-  filterAttributesByType,
-  getUniqueTypes,
-} from '@/lib/attribute-hooks'
-import { NestedCategory, nestCategories } from '@/lib/category-hierarchy'
+import { nestCategories } from '@/lib/category-hierarchy'
 
 import {
   Form,
@@ -28,9 +23,11 @@ import {
   FormMessage,
 } from '@/components/ui/form'
 
+import { CategoryList } from './category-list'
+import { AttributeList } from './attribute-list'
+
 import { Badge } from '@/components/ui/badge'
 import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
 import { Button } from '@/components/ui/button'
 import { Textarea } from '@/components/ui/textarea'
 import { Checkbox } from '@/components/ui/checkbox'
@@ -40,97 +37,51 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 
 import { AlertTriangle, ArrowRight, PlusCircle } from 'lucide-react'
 
-import { Large, Small } from '@/components/typography/texts'
-
 import { createProduct } from '../_logic/actions'
-
-const newProductSchema = z.object({
-  // Non-optional fields
-  title: z.string().min(3, 'Título do produto deve conter no mínimo 3 letras.'),
-  active: z.boolean(),
-  featuredImage: z.string(),
-  minimumQuantity: z.coerce
-    .number({
-      required_error: 'Campo deve ser preenchido.',
-      invalid_type_error: 'Campo deve conter um número.',
-    })
-    .min(1, 'Quantidade mínima deve ser maior que 1 unidade.'),
-  sku: z
-    .string()
-    .refine(
-      (value) => /^[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?$/.test(value),
-      'SKU deve ser alfanumérico e pode conter hífens, porém não no início nem no fim do código.',
-    ),
-
-  // Optional fields
-  description: z
-    .string()
-    .max(300, 'Uma descrição pode ter no máximo 300 caracteres.')
-    .optional(),
-
-  // Relationship fields
-  attributes: z.array(z.string()).optional(),
-  categories: z.array(z.string()).optional(),
-})
+import { newProductSchema } from '../_logic/validations'
 
 interface NewProductFormProps {
   setOpen: (state: boolean) => void
 }
 
 export function NewProductForm({ setOpen }: NewProductFormProps) {
-  const [attributes, setAttributes] = useState<Attribute[]>([])
-  const [categories, setCategories] = useState<Category[]>([])
   const [isLoading, setLoading] = useState<boolean>(false)
+  const [categories, setCategories] = useState<Category[]>([])
+  const [attributes, setAttributes] = useState<Attribute[]>([])
+
   const [activeTab, setActiveTab] = useState<string>('product')
 
   function handleChangeStep(destination: string) {
     setActiveTab(destination)
   }
 
-  const form = useForm<z.infer<typeof newProductSchema>>({
-    resolver: zodResolver(newProductSchema),
-    defaultValues: {
-      sku: '',
-      title: '',
-      description: '',
-      minimumQuantity: 0,
-      featuredImage: '660315294d562058e2daa8fb',
-      active: false,
-
-      categories: [],
-      attributes: [],
-    },
-  })
-
   useEffect(() => {
-    const fetchAttributes = async () => {
-      try {
-        setLoading(true)
-        const response = await fetch(`/api/all-attributes`)
+    async function fetchAttributes() {
+      setLoading(true)
 
+      try {
+        const response = await fetch(`/api/all-attributes`)
         const data = await response.json()
 
         setAttributes(data)
-        setLoading(false)
       } catch (error) {
-        console.error(error)
         toast.error(error)
+      } finally {
         setLoading(false)
       }
     }
 
-    const fetchCategories = async () => {
-      try {
-        setLoading(true)
-        const response = await fetch(`/api/all-categories`)
+    async function fetchCategories() {
+      setLoading(true)
 
+      try {
+        const response = await fetch(`/api/all-categories`)
         const data = await response.json()
 
         setCategories(data)
-        setLoading(false)
       } catch (error) {
-        console.error(error)
         toast.error(error)
+      } finally {
         setLoading(false)
       }
     }
@@ -139,7 +90,23 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
     fetchAttributes()
   }, [])
 
-  // Values is already formatted and validated.
+  const form = useForm<z.infer<typeof newProductSchema>>({
+    resolver: zodResolver(newProductSchema),
+    defaultValues: {
+      sku: '',
+      title: '',
+      description: '',
+      minimumQuantity: 0,
+      active: false,
+      featuredImage: new File([], ''),
+
+      categories: [],
+      attributes: [],
+    },
+  })
+
+  const { isSubmitting } = useFormState({ control: form.control })
+
   async function onSubmit(values: z.infer<typeof newProductSchema>) {
     const {
       sku,
@@ -152,6 +119,8 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
       categories,
     } = values
 
+    console.log('FeaturedImage onSubmit', JSON.stringify(featuredImage))
+
     const response = await createProduct({
       sku,
       title,
@@ -160,7 +129,7 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
       active,
 
       // Placeholder image id
-      featuredImage,
+      featuredImage: 'test',
 
       attributes,
       categories,
@@ -177,163 +146,6 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
 
       setOpen(false)
     }
-  }
-
-  interface AttributeListProps {
-    attributes: Attribute[]
-    props: any
-  }
-
-  const AttributeList: React.FC<AttributeListProps> = ({
-    attributes,
-    props,
-  }) => {
-    const colors = filterAttributesByType(attributes, 'color')
-    const labels = filterAttributesByType(attributes, 'label')
-    const types = getUniqueTypes(labels)
-
-    function handleCheckedChange(id: string, state: boolean) {
-      // state === true => Checkbox foi marcada => Atualiza estado do form adicionando o elemento
-      // state === false => Checkbox foi desmarcada => Remove o elemento do valor do form
-      state
-        ? props.value.push(id)
-        : (props.value = props.value.filter(
-            (attributeId: any) => attributeId !== id,
-          ))
-
-      // Pra funcionar com form controlado
-      form.setValue('attributes', props.value)
-    }
-
-    return (
-      <motion.div
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        className='grid w-full grid-cols-2 gap-4'
-      >
-        <div className='space-y-2.5'>
-          <Large>Cores</Large>
-
-          {colors.map((color) => (
-            <div key={color.id} className='flex items-center gap-1'>
-              <Checkbox
-                id={color.id}
-                name={color.name}
-                value={color.id}
-                defaultChecked={props.value.includes(color.id)}
-                onCheckedChange={(state) =>
-                  handleCheckedChange(color.id, !!state)
-                }
-              />
-              <div
-                className='h-5 w-5 rounded-full border'
-                style={{ backgroundColor: color.value }}
-              />
-
-              <Small>{color.name}</Small>
-            </div>
-          ))}
-        </div>
-
-        {types.map((type) => {
-          const filteredAttributes = filterAttributesByName(labels, type)
-
-          return (
-            <div key={type} className='space-y-1'>
-              <Large>{type}</Large>
-
-              {filteredAttributes.map((label) => (
-                <div key={label.id} className='flex items-center gap-1'>
-                  <Checkbox
-                    id={label.id}
-                    name={label.name}
-                    value={label.id}
-                    defaultChecked={props.value.includes(label.id)}
-                    onCheckedChange={(state) =>
-                      handleCheckedChange(label.id, !!state)
-                    }
-                  />
-
-                  <Small>{label.name}</Small>
-                </div>
-              ))}
-            </div>
-          )
-        })}
-      </motion.div>
-    )
-  }
-
-  interface CategoryListProps {
-    categories: NestedCategory[]
-    props: any
-  }
-
-  const CategoryList: React.FC<CategoryListProps> = ({ categories, props }) => {
-    return (
-      <div className='space-y-2'>
-        {categories.map((category) => (
-          <CategoryCheckbox
-            key={category.id}
-            category={category}
-            props={props}
-          />
-        ))}
-      </div>
-    )
-  }
-
-  interface CategoryCheckboxProps {
-    category: NestedCategory
-    props: any
-  }
-
-  const CategoryCheckbox: React.FC<CategoryCheckboxProps> = ({
-    category,
-    props,
-  }) => {
-    function handleCheckedChange(id: string, state: boolean) {
-      // state === true => Checkbox foi marcada => Atualiza estado do form adicionando o elemento
-      // state === false => Checkbox foi desmarcada => Remove o elemento do valor do form
-      state
-        ? props.value.push(id)
-        : (props.value = props.value.filter(
-            (categoryId: any) => categoryId !== id,
-          ))
-
-      // Pra funcionar com form controlado
-      form.setValue('categories', props.value)
-    }
-
-    return (
-      <div className='ml-4 space-y-2'>
-        <div className='group flex items-center gap-1.5'>
-          <Checkbox
-            id={category.id}
-            name={category.title}
-            value={category.id}
-            defaultChecked={props.value.includes(category.id)}
-            onCheckedChange={(state) =>
-              handleCheckedChange(category.id, !!state)
-            }
-          />
-          <Label
-            className='cursor-pointer hover:underline group-hover:underline'
-            htmlFor={category.id}
-          >
-            {category.title}
-          </Label>
-        </div>
-
-        {category.children.map((childCategory) => (
-          <CategoryCheckbox
-            key={childCategory.id}
-            category={childCategory}
-            props={props}
-          />
-        ))}
-      </div>
-    )
   }
 
   return (
@@ -460,10 +272,19 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
 
                           <FormControl>
                             {/* <Input id='featured' type='file' {...field} /> */}
-                            <Input id='featured' type='text' {...field} />
+                            <Input
+                              type='file'
+                              onChange={(e) =>
+                                field.onChange(
+                                  e.target.files ? e.target.files[0] : null,
+                                )
+                              }
+                            />
                           </FormControl>
+
                           <FormDescription>
-                            Escolha uma imagem de até 999x777 px
+                            Escolha uma imagem de até 999x777 px - png, jpg ou
+                            jpeg
                           </FormDescription>
                           <StyledFormMessage
                             form={form}
@@ -535,21 +356,30 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
                                   exit={{ opacity: 0 }}
                                   className='inline-flex flex-wrap gap-1 rounded-md'
                                 >
-                                  {field.value.map((item: string) => (
-                                    <Badge key={item} className='h-fit w-fit'>
-                                      {
-                                        attributes.find(
-                                          (attribute) => attribute.id === item,
-                                        ).name
-                                      }
-                                    </Badge>
-                                  ))}
+                                  {field.value.map((item: string) => {
+                                    const attribute = attributes.find(
+                                      (attribute) => attribute.id === item,
+                                    )
+
+                                    return (
+                                      <Badge key={item} className='h-fit w-fit'>
+                                        {attribute.name}
+                                      </Badge>
+                                    )
+                                  })}
                                 </motion.div>
                               )}
                             </AnimatePresence>
 
+                            {/* Render AttributeList, else render LoadingSpinner */}
                             <AnimatePresence>
-                              {isLoading && (
+                              {attributes && !isLoading ? (
+                                <AttributeList
+                                  attributes={attributes}
+                                  field={field}
+                                  set={form.setValue}
+                                />
+                              ) : (
                                 <motion.div
                                   initial={{ opacity: 0 }}
                                   animate={{ opacity: 1 }}
@@ -558,15 +388,6 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
                                 >
                                   <LoadingSpinner />
                                 </motion.div>
-                              )}
-                            </AnimatePresence>
-
-                            <AnimatePresence>
-                              {attributes && !isLoading && (
-                                <AttributeList
-                                  attributes={attributes}
-                                  props={field}
-                                />
                               )}
                             </AnimatePresence>
                             <FormMessage />
@@ -611,15 +432,17 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
                                   exit={{ opacity: 0 }}
                                   className='inline-flex flex-wrap gap-1 rounded-md'
                                 >
-                                  {field.value.map((item: string) => (
-                                    <Badge key={item} className='h-fit w-fit'>
-                                      {
-                                        categories.find(
-                                          (category) => category.id === item,
-                                        ).title
-                                      }
-                                    </Badge>
-                                  ))}
+                                  {field.value.map((item: string) => {
+                                    const category = categories.find(
+                                      (category) => category.id === item,
+                                    )
+
+                                    return (
+                                      <Badge key={item} className='h-fit w-fit'>
+                                        {category.title}
+                                      </Badge>
+                                    )
+                                  })}
                                 </motion.div>
                               )}
                             </AnimatePresence>
@@ -640,7 +463,8 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
                             {categories && (
                               <CategoryList
                                 categories={nestCategories(categories)}
-                                props={field}
+                                field={field}
+                                set={form.setValue}
                               />
                             )}
 
@@ -650,7 +474,11 @@ export function NewProductForm({ setOpen }: NewProductFormProps) {
                       />
                     </ScrollArea>
 
-                    <Button type='submit' className='col-span-2 w-full'>
+                    <Button
+                      type='submit'
+                      disabled={isSubmitting}
+                      className='col-span-2 w-full'
+                    >
                       <PlusCircle className='mr-2 h-5 w-5' />
                       Criar produto
                     </Button>
