+++ src/app/(dashboard)/dashboard/catalog/produtos/_components/update-product-dialog-content.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import { Product } from '@/payload/payload-types'
 
 import {
