+++ src/app/(dashboard)/dashboard/catalog/produtos/_components/update-product-form.tsx
@@ -3,15 +3,10 @@
 import { useEffect, useState } from 'react'
 
 import { z } from 'zod'
-import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
+import { useForm, useFormState } from 'react-hook-form'
 
-import {
-  filterAttributesByName,
-  filterAttributesByType,
-  getUniqueTypes,
-} from '@/lib/attribute-hooks'
-import { NestedCategory, nestCategories } from '@/lib/category-hierarchy'
+import { nestCategories } from '@/lib/category-hierarchy'
 
 import { toast } from 'sonner'
 import { AnimatePresence, motion } from 'framer-motion'
@@ -28,7 +23,9 @@ import {
   FormMessage,
 } from '@/components/ui/form'
 
-import { Label } from '@/components/ui/label'
+import { CategoryList } from './category-list'
+import { AttributeList } from './attribute-list'
+
 import { Input } from '@/components/ui/input'
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
@@ -40,38 +37,8 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 
 import { AlertTriangle, ArrowRight, Pencil } from 'lucide-react'
 
-import { Large, Small } from '@/components/typography/texts'
-
 import { updateProduct } from '../_logic/actions'
-
-const updateProductSchema = z.object({
-  // Non-optional fields
-  title: z.string().min(3, 'Título do produto deve conter no mínimo 3 letras.'),
-  active: z.boolean(),
-  featuredImage: z.string(),
-  minimumQuantity: z.coerce
-    .number({
-      required_error: 'Campo deve ser preenchido.',
-      invalid_type_error: 'Campo deve conter um número.',
-    })
-    .min(1, 'Quantidade mínima deve ser maior que 1 unidade.'),
-  sku: z
-    .string()
-    .refine(
-      (value) => /^[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?$/.test(value),
-      'SKU deve ser alfanumérico e pode conter hífens, porém não no início nem no fim do código.',
-    ),
-
-  // Optional fields
-  description: z
-    .string()
-    .max(300, 'Uma descrição pode ter no máximo 300 caracteres.')
-    .optional(),
-
-  // Relationship fields
-  attributes: z.array(z.string()).optional(),
-  categories: z.array(z.string()).optional(),
-})
+import { updateProductSchema } from '../_logic/validations'
 
 interface UpdateProductFormProps {
   currentProduct: Product
@@ -82,6 +49,51 @@ export function UpdateProductForm({
   currentProduct,
   setOpen,
 }: UpdateProductFormProps) {
+  const [isLoading, setLoading] = useState<boolean>(false)
+  const [categories, setCategories] = useState<Category[]>([])
+  const [attributes, setAttributes] = useState<Attribute[]>([])
+
+  const [activeTab, setActiveTab] = useState<string>('product')
+
+  function handleChangeStep(destination: string) {
+    setActiveTab(destination)
+  }
+
+  useEffect(() => {
+    async function fetchAttributes() {
+      setLoading(true)
+
+      try {
+        const response = await fetch(`/api/all-attributes`)
+        const data = await response.json()
+
+        setAttributes(data)
+      } catch (error) {
+        toast.error(error)
+      } finally {
+        setLoading(false)
+      }
+    }
+
+    async function fetchCategories() {
+      setLoading(true)
+
+      try {
+        const response = await fetch(`/api/all-categories`)
+        const data = await response.json()
+
+        setCategories(data)
+      } catch (error) {
+        toast.error(error)
+      } finally {
+        setLoading(false)
+      }
+    }
+
+    fetchCategories()
+    fetchAttributes()
+  }, [])
+
   const form = useForm<z.infer<typeof updateProductSchema>>({
     resolver: zodResolver(updateProductSchema),
     defaultValues: {
@@ -108,205 +120,8 @@ export function UpdateProductForm({
     },
   })
 
-  const [attributes, setAttributes] = useState<Attribute[]>([])
-  const [categories, setCategories] = useState<Category[]>([])
-  const [isLoading, setLoading] = useState<boolean>(false)
+  const { isSubmitting } = useFormState({ control: form.control })
 
-  useEffect(() => {
-    const fetchAttributes = async () => {
-      try {
-        setLoading(true)
-        const response = await fetch(`/api/all-attributes`)
-
-        const data = await response.json()
-
-        setAttributes(data)
-        setLoading(false)
-      } catch (error) {
-        console.error(error)
-        toast.error(error)
-        setLoading(false)
-      }
-    }
-
-    const fetchCategories = async () => {
-      try {
-        setLoading(true)
-        const response = await fetch(`/api/all-categories`)
-
-        const data = await response.json()
-
-        setCategories(data)
-        setLoading(false)
-      } catch (error) {
-        console.error(error)
-        toast.error(error)
-        setLoading(false)
-      }
-    }
-
-    fetchCategories()
-    fetchAttributes()
-  }, [])
-
-  interface AttributeListProps {
-    attributes: Attribute[]
-    props: any
-  }
-
-  const AttributeList: React.FC<AttributeListProps> = ({
-    attributes,
-    props,
-  }) => {
-    const colors = filterAttributesByType(attributes, 'color')
-    const labels = filterAttributesByType(attributes, 'label')
-    const types = getUniqueTypes(labels)
-
-    function handleCheckedChange(id: string, state: boolean) {
-      // state === true => Checkbox foi marcada => Atualiza estado do form adicionando o elemento
-      // state === false => Checkbox foi desmarcada => Remove o elemento do valor do form
-      state
-        ? props.value.push(id)
-        : (props.value = props.value.filter(
-            (attributeId: any) => attributeId !== id,
-          ))
-
-      // Pra funcionar com form controlado
-      form.setValue('attributes', props.value)
-    }
-
-    return (
-      <motion.div
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        className='grid w-full grid-cols-2 gap-4'
-      >
-        <div className='space-y-2.5'>
-          <Large>Cores</Large>
-
-          {colors.map((color) => (
-            <div key={color.id} className='flex items-center gap-1'>
-              <Checkbox
-                id={color.id}
-                name={color.name}
-                value={color.id}
-                defaultChecked={props.value.includes(color.id)}
-                onCheckedChange={(state) =>
-                  handleCheckedChange(color.id, !!state)
-                }
-              />
-              <div
-                className='h-5 w-5 rounded-full border'
-                style={{ backgroundColor: color.value }}
-              />
-
-              <Small>{color.name}</Small>
-            </div>
-          ))}
-        </div>
-
-        {types.map((type) => {
-          const filteredAttributes = filterAttributesByName(labels, type)
-
-          return (
-            <div key={type} className='space-y-1'>
-              <Large>{type}</Large>
-
-              {filteredAttributes.map((label) => (
-                <div key={label.id} className='flex items-center gap-1'>
-                  <Checkbox
-                    id={label.id}
-                    name={label.name}
-                    value={label.id}
-                    defaultChecked={props.value.includes(label.id)}
-                    onCheckedChange={(state) =>
-                      handleCheckedChange(label.id, !!state)
-                    }
-                  />
-
-                  <Small>{label.name}</Small>
-                </div>
-              ))}
-            </div>
-          )
-        })}
-      </motion.div>
-    )
-  }
-
-  interface CategoryListProps {
-    categories: NestedCategory[]
-    props: any
-  }
-
-  const CategoryList: React.FC<CategoryListProps> = ({ categories, props }) => {
-    return (
-      <div className='space-y-2'>
-        {categories.map((category) => (
-          <CategoryCheckbox
-            key={category.id}
-            category={category}
-            props={props}
-          />
-        ))}
-      </div>
-    )
-  }
-
-  interface CategoryCheckboxProps {
-    category: NestedCategory
-    props: any
-  }
-
-  const CategoryCheckbox: React.FC<CategoryCheckboxProps> = ({
-    category,
-    props,
-  }) => {
-    function handleCheckedChange(id: string, state: boolean) {
-      // state === true => Checkbox foi marcada => Atualiza estado do form adicionando o elemento
-      // state === false => Checkbox foi desmarcada => Remove o elemento do valor do form
-      state
-        ? props.value.push(id)
-        : (props.value = props.value.filter(
-            (categoryId: any) => categoryId !== id,
-          ))
-
-      // Pra funcionar com form controlado
-      form.setValue('categories', props.value)
-    }
-
-    return (
-      <div className='ml-4 space-y-2'>
-        <div className='group flex items-center gap-1.5'>
-          <Checkbox
-            id={category.id}
-            name={category.title}
-            value={category.id}
-            defaultChecked={props.value.includes(category.id)}
-            onCheckedChange={(state) =>
-              handleCheckedChange(category.id, !!state)
-            }
-          />
-          <Label
-            className='cursor-pointer hover:underline group-hover:underline'
-            htmlFor={category.id}
-          >
-            {category.title}
-          </Label>
-        </div>
-
-        {category.children.map((childCategory) => (
-          <CategoryCheckbox
-            key={childCategory.id}
-            category={childCategory}
-            props={props}
-          />
-        ))}
-      </div>
-    )
-  }
-
-  // Values is already formatted and validated.
   async function onSubmit(values: z.infer<typeof updateProductSchema>) {
     const {
       sku,
@@ -345,12 +160,6 @@ export function UpdateProductForm({
       setOpen(false)
     }
   }
-
-  function handleChangeStep(destination: string) {
-    setActiveTab(destination)
-  }
-
-  const [activeTab, setActiveTab] = useState<string>('product')
 
   return (
     <Form {...form}>
@@ -559,21 +368,30 @@ export function UpdateProductForm({
                                   exit={{ opacity: 0 }}
                                   className='inline-flex flex-wrap gap-1 rounded-md'
                                 >
-                                  {field.value.map((item: string) => (
-                                    <Badge key={item} className='h-fit w-fit'>
-                                      {
-                                        attributes.find(
-                                          (attribute) => attribute.id === item,
-                                        ).name
-                                      }
-                                    </Badge>
-                                  ))}
+                                  {field.value.map((item: string) => {
+                                    const attribute = attributes.find(
+                                      (attribute) => attribute.id === item,
+                                    )
+
+                                    return (
+                                      <Badge key={item} className='h-fit w-fit'>
+                                        {attribute.name}
+                                      </Badge>
+                                    )
+                                  })}
                                 </motion.div>
                               )}
                             </AnimatePresence>
 
+                            {/* Render AttributeList, else render LoadingSpinner */}
                             <AnimatePresence>
-                              {isLoading && (
+                              {attributes && !isLoading ? (
+                                <AttributeList
+                                  attributes={attributes}
+                                  field={field}
+                                  set={form.setValue}
+                                />
+                              ) : (
                                 <motion.div
                                   initial={{ opacity: 0 }}
                                   animate={{ opacity: 1 }}
@@ -582,15 +400,6 @@ export function UpdateProductForm({
                                 >
                                   <LoadingSpinner />
                                 </motion.div>
-                              )}
-                            </AnimatePresence>
-
-                            <AnimatePresence>
-                              {attributes && !isLoading && (
-                                <AttributeList
-                                  attributes={attributes}
-                                  props={field}
-                                />
                               )}
                             </AnimatePresence>
                             <FormMessage />
@@ -635,15 +444,17 @@ export function UpdateProductForm({
                                   exit={{ opacity: 0 }}
                                   className='inline-flex flex-wrap gap-1 rounded-md'
                                 >
-                                  {field.value.map((item: string) => (
-                                    <Badge key={item} className='h-fit w-fit'>
-                                      {
-                                        categories.find(
-                                          (category) => category.id === item,
-                                        ).title
-                                      }
-                                    </Badge>
-                                  ))}
+                                  {field.value.map((item: string) => {
+                                    const category = categories.find(
+                                      (category) => category.id === item,
+                                    )
+
+                                    return (
+                                      <Badge key={item} className='h-fit w-fit'>
+                                        {category.title}
+                                      </Badge>
+                                    )
+                                  })}
                                 </motion.div>
                               )}
                             </AnimatePresence>
@@ -651,7 +462,8 @@ export function UpdateProductForm({
                             {categories && (
                               <CategoryList
                                 categories={nestCategories(categories)}
-                                props={field}
+                                field={field}
+                                set={form.setValue}
                               />
                             )}
 
@@ -661,7 +473,11 @@ export function UpdateProductForm({
                       />
                     </ScrollArea>
 
-                    <Button type='submit' className='col-span-2 w-full'>
+                    <Button
+                      type='submit'
+                      disabled={isSubmitting}
+                      className='col-span-2 w-full'
+                    >
                       <Pencil className='mr-2 h-5 w-5' />
                       Atualizar produto
                     </Button>
